+++ pyfredapi/api/__init__.py
@@ -3,5 +3,6 @@ from .category import FredCategory
 from .maps import FredMaps
 from .releases import FredRelease
 from .series import FredSeries
+from .series_collection import SeriesCollection
 from .sources import FredSources
 from .tags import FredTags