+++ pyfredapi/api/releases.py
@@ -1,3 +1,5 @@
+"""This module contains the FredRelease implementation."""
+
 from typing import Literal, Optional
 
 from pydantic import BaseModel, PositiveInt